Clean up data dragon api stuff

+++ backend/src/main/kotlin/in/aika/riotdataexplorer/api/model/general/spectator/BannedChampion.kt
@@ -4,6 +4,6 @@ import `in`.aika.riotdataexplorer.api.model.LogIgnoredProperties
 
 data class BannedChampion(
     val pickTurn: Int,
-    val championId: Long,
+    val championId: Int,
     val teamId: Long,
 ) : LogIgnoredProperties()
